Fix global posting list loading bug and lock handling

+++ mailpile/postinglist.py
@@ -1,6 +1,7 @@
 import os
 import random
 import threading
+import time
 from gettext import gettext as _
 
 import mailpile.util
@@ -44,13 +45,10 @@ class PostingList(object):
                 if force or (filesize > 900 * postinglist_kb):
                     session.ui.mark('Pass 1: Compacting >%s<' % fn)
                     play_nice_with_threads()
-                    try:
-                        GLOBAL_POSTING_LOCK.acquire()
+                    with GLOBAL_POSTING_LOCK:
                         # FIXME: Remove invalid and deleted messages from
                         #        posting lists.
                         cls(session, fn, sig=fn).save()
-                    finally:
-                        GLOBAL_POSTING_LOCK.release()
 
         # Pass 2: While mergable pair exists: merge them!
         for c in cls.CHARACTERS:
@@ -65,15 +63,18 @@ class PostingList(object):
                 while not os.path.exists(os.path.join(postinglist_dir, fnp)):
                     fnp = fnp[:-1]
                 size += os.path.getsize(os.path.join(postinglist_dir, fnp))
-                if (size < (1024 * postinglist_kb - (cls.HASH_LEN * 6))):
+                if (fnp and
+                    size < (1024 * postinglist_kb - (cls.HASH_LEN * 6))):
                     session.ui.mark('Pass 2: Merging %s into %s' % (fn, fnp))
                     play_nice_with_threads()
                     try:
                         GLOBAL_POSTING_LOCK.acquire()
-                        with open(os.path.join(postinglist_dir, fn), 'r') as fd, \
-                                open(os.path.join(postinglist_dir, fnp), 'a') as fdp:
-                            for line in fd:
-                                fdp.write(line)
+                        path_fn = os.path.join(postinglist_dir, fn)
+                        path_fnp = os.path.join(postinglist_dir, fnp)
+                        with open(path_fn, 'r') as fd:
+                            with open(path_fnp, 'a') as fdp:
+                                for line in fd:
+                                    fdp.write(line)
                     finally:
                         os.remove(os.path.join(postinglist_dir, fn))
                         GLOBAL_POSTING_LOCK.release()
@@ -88,34 +89,39 @@ class PostingList(object):
     def _Append(cls, session, word, mail_ids, compact=True, sig=None):
         config = session.config
         sig = sig or cls.WordSig(word, config)
-        fd, fn = cls.GetFile(session, sig, mode='a')
-        try:
-            if (compact
-                    and (os.path.getsize(os.path.join(config.postinglist_dir(fn),
-                         fn)) > ((1024 * config.sys.postinglist_kb) -
-                                 (cls.HASH_LEN * 6)))
-                    and (random.randint(0, 50) == 1)):
-                # This will compact the files and split out hot-spots, but we
-                # only bother "once in a while" when the files are "big".
-                fd.close()
-                pls = cls(session, word, sig=sig)
-                for mail_id in mail_ids:
-                    pls.append(mail_id)
-                pls.save()
-            else:
-                # Quick and dirty append is the default.
-                fd.write('%s\t%s\n' % (sig, '\t'.join(mail_ids)))
-        finally:
-            if not fd.closed:
-                fd.close()
+
+        fd = None
+        while fd is None:
+            fd, fn = cls.GetFile(session, sig, mode='a')
+            fn_path = cls.SaveFile(session, fn)
+            try:
+                # The code below will compact the files and split out hot-spots,
+                # but we only bother "once in a while" when the files are "big".
+                if compact:
+                    max_size = ((1024 * config.sys.postinglist_kb) -
+                                (cls.HASH_LEN * 6))
+                    compact = (os.path.getsize(fn_path) > max_size and
+                               random.randint(0, 50) == 1)
+                if fd:
+                    with fd:
+                        if not compact:
+                            fd.write('%s\t%s\n' % (sig, '\t'.join(mail_ids)))
+                            return
+            except IOError:
+                print 'OMGWTF: compact=%s %s / %s' % (compact, fn_path, fd)
+                time.sleep(0.2)
+                fd = None
+
+        # OK, compactinate!
+        pls = cls(session, word, sig=sig)
+        for mail_id in mail_ids:
+            pls.append(mail_id)
+        pls.save()
 
     @classmethod
     def Lock(cls, lock, method, *args, **kwargs):
-        lock.acquire()
-        try:
+        with lock:
             return method(*args, **kwargs)
-        finally:
-            lock.release()
 
     @classmethod
     def Optimize(cls, *args, **kwargs):
@@ -185,8 +191,8 @@ class PostingList(object):
             except ValueError:
                 pass
             finally:
-                fd.close()
                 self.lock.release()
+                fd.close()
 
     def _fmt_file(self, prefix):
         output = []
@@ -244,26 +250,20 @@ class PostingList(object):
         return self.WORDS[self.sig]
 
     def append(self, eid):
-        self.lock.acquire()
-        try:
+        with self.lock:
             if self.sig not in self.WORDS:
                 self.WORDS[self.sig] = set()
             self.WORDS[self.sig].add(eid)
             return self
-        finally:
-            self.lock.release()
 
     def remove(self, eids):
-        self.lock.acquire()
-        try:
+        with self.lock:
             for eid in eids:
                 try:
                     self.WORDS[self.sig].remove(eid)
                 except KeyError:
                     pass
             return self
-        finally:
-            self.lock.release()
 
 
 GLOBAL_GPL_LOCK = threading.Lock()
@@ -316,8 +316,7 @@ class GlobalPostingList(PostingList):
         super(GlobalPostingList, cls)._Append(session, word, mail_ids,
                                               compact=compact)
         global GLOBAL_POSTING_LIST
-        GLOBAL_GPL_LOCK.acquire()
-        try:
+        with GLOBAL_GPL_LOCK:
             sig = cls.WordSig(word, session.config)
             if GLOBAL_POSTING_LIST is None:
                 GLOBAL_POSTING_LIST = {}
@@ -325,8 +324,6 @@ class GlobalPostingList(PostingList):
                 GLOBAL_POSTING_LIST[sig] = set()
             for mail_id in mail_ids:
                 GLOBAL_POSTING_LIST[sig].add(mail_id)
-        finally:
-            GLOBAL_GPL_LOCK.release()
 
     def __init__(self, *args, **kwargs):
         PostingList.__init__(self, *args, **kwargs)
@@ -339,24 +336,22 @@ class GlobalPostingList(PostingList):
         return prefix, output
 
     def load(self):
-        self.filename = 'kw-journal.dat'
-        global GLOBAL_POSTING_LIST
-        if GLOBAL_POSTING_LIST:
-            self.WORDS = GLOBAL_POSTING_LIST
-        else:
-            PostingList.load(self)
-            GLOBAL_POSTING_LIST = self.WORDS
+        with GLOBAL_GPL_LOCK:
+            self.filename = 'kw-journal.dat'
+            global GLOBAL_POSTING_LIST
+            if GLOBAL_POSTING_LIST:
+                self.WORDS = GLOBAL_POSTING_LIST
+            else:
+                PostingList.load(self)
+                GLOBAL_POSTING_LIST = self.WORDS
 
     def _migrate(self, sig=None, compact=True):
-        self.lock.acquire()
-        try:
+        with self.lock:
             sig = sig or self.sig
             if sig in self.WORDS and len(self.WORDS[sig]) > 0:
                 PostingList.Append(self.session, sig, self.WORDS[sig],
                                    sig=sig, compact=compact)
                 del self.WORDS[sig]
-        finally:
-            self.lock.release()
 
     def remove(self, eids):
         PostingList(self.session, self.word,
